Added Secure lines to the views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Product, Category
 from profiles.models import UserProfile
@@ -9,7 +10,6 @@ from .forms import ProductForm
 
 def all_products(request):
 
-    
     products = Product.objects.all()
     query = None
     categories = None
@@ -67,8 +67,14 @@ def product_detail(request, product_id):
     }
     return render(request, 'products/product_detail.html', context)
 
+
+@login_required
 def add_product(request):
-    """ Add a product to the store """
+    """ Add a product to the store only by superuser """
+    if not request.user.is_superuser:
+        messages.error(request, 'Error, Unauthorized User')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -88,9 +94,13 @@ def add_product(request):
     
     return render(request, template, context)
 
-
+@login_required
 def edit_product(request, product_id):
-    """ Edit a product in the store """
+    """ Edit a product in the store only by superuser """
+    if not request.user.is_superuser:
+        messages.error(request, 'Error, Unauthorized User.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -113,9 +123,13 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_product(request, product_id):
-    """ Delete a product from the store """
+    """ Delete a product from the store only by superuser """
+    if not request.user.is_superuser:
+        messages.error(request, 'Error, Unauthorized User.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
